Test new riff system (#43)
The runtimes have been deconstructed into `riff-core-runtime` and
`riff-knative-runtime` charts. There is no longer a central `riff`
chart, although we may bring it back.

Removing kpack bits that are not quite ready, although they will also
come back in the future.

+++ ci/cleanup-riff.sh
@@ -1,15 +1,33 @@
 #!/bin/bash
 
-tiller_service_account=${1:-tiller}
-tiller_namespace=${2:-kube-system}
+tiller_service_account=tiller
+tiller_namespace=kube-system
 
-helm delete --purge riff
-kubectl delete customresourcedefinitions.apiextensions.k8s.io -l app.kubernetes.io/managed-by=Tiller,app.kubernetes.io/instance=riff
+uninstall_chart() {
+  local name=$1
 
-helm delete --purge istio
-kubectl delete customresourcedefinitions.apiextensions.k8s.io -l app.kubernetes.io/managed-by=Tiller,app.kubernetes.io/instance=istio
-kubectl delete namespace istio-system
+  helm delete --purge $name
+  kubectl delete customresourcedefinitions.apiextensions.k8s.io -l app.kubernetes.io/managed-by=Tiller,app.kubernetes.io/instance=$name 
+}
 
+if [ $RUNTIME = "core" ]; then
+  echo "Uninstall riff Core Runtime"
+  uninstall_chart riff-core-runtime
+
+elif [ $RUNTIME = "knative" ]; then
+  echo "Uninstall riff Knative Runtime"
+  uninstall_chart riff-knative-runtime
+
+  uninstall_chart istio
+  # extra cleanup for Istio
+  kubectl get customresourcedefinitions.apiextensions.k8s.io -oname | grep istio.io | xargs -L1 kubectl delete
+  kubectl delete namespace istio-system
+fi
+
+echo "Uninstall riff Build"
+uninstall_chart riff-build
+
+echo "Uninstall helm"
 helm reset
 
 kubectl delete serviceaccount ${tiller_service_account} -n ${tiller_namespace}
